passphrase cleanup
