--- app/components/modals/PassphraseModal/index.js
@@ -1,4 +1,0 @@
-export { default as PassphraseModal } from "./PassphraseModal";
-export { default as Modal } from "./PassphraseModalContent";
-export { default as ButtonsToolbar } from "./PassphraseModalButtonsToolbar";
-export { default as PassphraseInputRow } from "./PassphraseModalInputRow";
